bug: fix generation
Problem: we currently cannot filter out weird emoji characters
that do not build.
Solution: I could not find a good way to do the filter via
regex, or even check in Python, so instead we do a jekyll
build at each addition, and if the exit code is nonzero we
remove the file. It takes much longer, but within the GHA
time limit so good for a first try.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>